requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.24.0
supabase==2.0.3
google-generativeai==0.3.2
openai==1.3.8
python-multipart==0.0.6
pydantic==2.5.0
aiofiles==23.2.1

